Development

+++ src/app/components/shared/comments/comments.component.ts
@@ -1,45 +1,33 @@
-import { Component, OnInit, Input } from '@angular/core';
+import { Component, OnInit, Input, ViewChild } from '@angular/core';
 import { trigger, animate, style, transition, animateChild, group, query, stagger } from '@angular/animations';
 import { Router } from '@angular/router';
 
-import {
-    TdBounceAnimation,
-    TdFlashAnimation,
-    TdHeadshakeAnimation,
-    TdJelloAnimation,
-    TdPulseAnimation,
-    TdFadeInOutAnimation
-} from '@covalent/core';
-
 import { CommentService } from 'app/services';
 import { CurrentUserService } from 'app/infrastructure/services';
-import { MediaViewModel, CurrentUserViewModel, CommentViewModel } from 'app/models/view';
+import { MediaViewModel, CurrentUserViewModel, CommentViewModel, Page, PaginationViewModel } from 'app/models/view';
 
 @Component({
     selector: 'app-comments',
     templateUrl: './comments.component.html',
     styleUrls: ['./comments.component.css'],
     animations: [
-        TdFadeInOutAnimation(),
-        TdBounceAnimation(),                    // using implicit anchor name 'tdBounce' in template
-        TdFlashAnimation(),                     // using implicit anchor name 'tdFlash' in template
-        TdHeadshakeAnimation(),                 // using implicit anchor name 'tdHeadshake' in template
-        TdJelloAnimation(),                     // using implicit anchor name 'tdJello' in template
-        TdPulseAnimation({ duration: 200 }),    // using implicit anchor name 'tdPulse' in template,
-        trigger('enterTransition', [
-            transition(':enter', [
-                style({ transform: 'translateX(25px)', opacity: 0 }),
-                animate('200ms cubic-bezier(0.35, 0, 0.25, 1)', style('*'))
+        trigger('commentsAnimation', [
+            transition('* => *', [
+                query('.comment', style({ transform: 'translateX(-50%)' })),
+                query('.comment',
+                    stagger('100ms', [
+                        animate('250ms', style({ transform: 'translateX(0)' }))
+                    ]))
             ])
         ])
     ]
 })
-export class CommentsComponent {
+export class CommentsComponent implements OnInit {
     @Input() public media: MediaViewModel;
     public isLoading: boolean;
-    public pulseState = false;
 
     private currentUser: CurrentUserViewModel;
+    public page: Page<CommentViewModel>;
 
     constructor(
         private currentUserService: CurrentUserService,
@@ -49,25 +37,59 @@ export class CommentsComponent {
     }
 
     public getComments() {
-        this.pulseState = !this.pulseState;
+        if (this.isLoading) {
+            return;
+        }
+
         this.isLoading = true;
 
-        this.commentService.getComments(this.media.id)
+        this.commentService.getComments(this.media.id, this.page.pagination)
             .finally(() => {
                 this.isLoading = false;
             })
-            .subscribe(comments => {
-                this.media.comments = comments;
+            .subscribe(page => {
+                this.page.hasMoreItems = page.hasMoreItems;
+                this.page.pagination = page.pagination;
+                if (page.data) {
+                    this.page.data = this.page.data.concat(page.data);
+                }
             });
     }
 
+    public createComment(comment: CommentViewModel) {
+        this.page.data.unshift(comment);
+    }
+
     public removeComment(comment: CommentViewModel) {
-        const indexToRemove = this.media.comments.findIndex(e => e.id === comment.id);
-        this.media.comments.splice(indexToRemove, 1);
-        this.media.commentsCount--;
-        this.commentService.removeComment(this.media.id, comment.id)
-            .subscribe(() => { }, () => {
-                this.media.commentsCount++;
-            });
+        if (comment.id) {
+            const indexToRemove = this.page.data.findIndex(e => e.id === comment.id);
+            this.page.data.splice(indexToRemove, 1);
+            this.media.commentsCount--;
+            this.commentService.removeComment(this.media.id, comment.id)
+                .subscribe(() => { }, () => {
+                    this.media.commentsCount++;
+                });
+        } else {
+            const indexToRemove = this.page.data.findIndex(e => !e.id);
+            this.page.data.splice(indexToRemove, 1);
+            this.media.commentsCount--;
+        }
+    }
+
+    public ngOnInit(): void {
+        this.page = new Page<CommentViewModel>();
+
+        if (this.page.data) {
+            this.page.hasMoreItems = this.media.comments.length < this.media.commentsCount;
+        }
+
+        if (this.page.hasMoreItems) {
+            this.page.pagination = new PaginationViewModel();
+            this.page.pagination.next = this.media.comments[this.media.comments.length - 1].id;
+
+            this.media.comments.splice(this.media.comments.length - 1, 1);
+        }
+
+        this.page.data = this.media.comments;
     }
 }
